refactor: 매장주문 Presentation Layer 분리
