--- src/main/java/kitchenpos/order/supports/strategy/OrderProcess.java
@@ -1,20 +1,0 @@
-package kitchenpos.order.supports.strategy;
-
-import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderType;
-
-public interface OrderProcess {
-    Order create(Order order);
-
-    Order accept(Order order);
-
-    Order serve(Order order);
-
-    Order complete(Order order);
-
-    Order startDelivery(Order order);
-
-    Order completeDelivery(Order order);
-
-    boolean support(OrderType orderType);
-}
