--- src/test/java/kitchenpos/orders/application/DeliveryOrderServiceTest.java
@@ -1,285 +1,0 @@
-package kitchenpos.orders.application;
-
-import kitchenpos.menus.application.InMemoryMenuRepository;
-import kitchenpos.menus.domain.MenuRepository;
-import kitchenpos.order.deliveryorders.application.DeliveryOrderService;
-import kitchenpos.order.deliveryorders.domain.*;
-import kitchenpos.order.domain.*;
-import kitchenpos.order.eatinorders.domain.OrderTableRepository;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.*;
-
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-import java.util.UUID;
-
-import static kitchenpos.Fixtures.*;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertAll;
-
-class DeliveryOrderServiceTest {
-    private OrderRepository orderRepository;
-    private MenuRepository menuRepository;
-    private OrderTableRepository orderTableRepository;
-    private FakeKitchenridersClient kitchenridersClient;
-    private OrderLineItemsValidService orderLineItemsValidService;
-
-    // DELIVERY order service
-    private DeliveryOrderService deliveryOrderService;
-    private DeliveryOrderCreateService deliveryOrderCreateService;
-    private DeliveryOrderAcceptService deliveryOrderAcceptService;
-    private DeliveryOrderServeService deliveryOrderServeService;
-    private DeliveryOrderCompleteService deliveryOrderCompleteService;
-    private DeliveryOrderStartDeliveryService deliveryOrderStartDeliveryService;
-    private DeliveryOrderCompleteDeliveryService deliveryOrderCompleteDeliveryService;
-
-
-    private static List<Arguments> orderLineItems() {
-        return Arrays.asList(
-                null,
-                Arguments.of(Collections.emptyList()),
-                Arguments.of(Arrays.asList(createOrderLineItemRequest(INVALID_ID, 19_000L, 3L)))
-        );
-    }
-
-    private static OrderLineItem createOrderLineItemRequest(final UUID menuId, final long price, final long quantity) {
-        return new OrderLineItem(menuId, quantity, BigDecimal.valueOf(price));
-    }
-
-    @BeforeEach
-    void setUp() {
-        orderRepository = new InMemoryOrderRepository();
-        menuRepository = new InMemoryMenuRepository();
-        orderTableRepository = new InMemoryOrderTableRepository();
-        kitchenridersClient = new FakeKitchenridersClient();
-        orderLineItemsValidService = new OrderLineItemsValidService(menuRepository);
-
-
-        deliveryOrderCreateService = new DeliveryOrderCreateService(orderRepository, orderLineItemsValidService);
-        deliveryOrderAcceptService = new DeliveryOrderAcceptService(orderRepository, kitchenridersClient);
-        deliveryOrderServeService = new DeliveryOrderServeService(orderRepository);
-        deliveryOrderCompleteService = new DeliveryOrderCompleteService(orderRepository);
-        deliveryOrderStartDeliveryService = new DeliveryOrderStartDeliveryService(orderRepository);
-        deliveryOrderCompleteDeliveryService = new DeliveryOrderCompleteDeliveryService(orderRepository);
-        deliveryOrderService = new DeliveryOrderService(orderRepository, deliveryOrderCreateService, deliveryOrderAcceptService, deliveryOrderServeService, deliveryOrderStartDeliveryService, deliveryOrderCompleteDeliveryService, deliveryOrderCompleteService);
-
-
-    }
-
-    @DisplayName("1개 이상의 등록된 메뉴로 배달 주문을 등록할 수 있다.")
-    @Test
-    void createDeliveryOrder() {
-        final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
-        final Order expected = createOrderRequest(
-                OrderType.DELIVERY, "서울시 송파구 위례성대로 2", createOrderLineItemRequest(menuId, 19_000L, 3L)
-        );
-        final Order actual = deliveryOrderService.create(expected);
-        assertThat(actual).isNotNull();
-        assertAll(
-                () -> assertThat(actual.getId()).isNotNull(),
-                () -> assertThat(actual.getType()).isEqualTo(expected.getType()),
-                () -> assertThat(actual.getStatus()).isEqualTo(OrderStatus.WAITING),
-                () -> assertThat(actual.getOrderDateTime()).isNotNull(),
-                () -> assertThat(actual.getOrderLineItems().getOrderLineItems()).hasSize(1),
-                () -> assertThat(actual.getDeliveryAddress()).isEqualTo(expected.getDeliveryAddress())
-        );
-    }
-
-
-    @DisplayName("주문 유형이 올바르지 않으면 등록할 수 없다.")
-    @NullSource
-    @ParameterizedTest
-    void create(final OrderType type) {
-        final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
-        final Order expected = createOrderRequest(type, createOrderLineItemRequest(menuId, 19_000L, 3L));
-        assertThatThrownBy(() -> deliveryOrderService.create(expected))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("메뉴가 없으면 등록할 수 없다.")
-    @MethodSource("orderLineItems")
-    @ParameterizedTest
-    void create(final List<OrderLineItem> orderLineItems) {
-        assertThatThrownBy(() -> {
-            final Order expected = createOrderRequest(OrderType.DELIVERY, orderLineItems);
-            deliveryOrderService.create(expected);
-        }).isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("배달 주소가 올바르지 않으면 배달 주문을 등록할 수 없다.")
-    @NullAndEmptySource
-    @ParameterizedTest
-    void create(final String deliveryAddress) {
-        final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
-        final Order expected = createOrderRequest(
-                OrderType.DELIVERY, deliveryAddress, createOrderLineItemRequest(menuId, 19_000L, 3L)
-        );
-        assertThatThrownBy(() -> deliveryOrderService.create(expected))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("숨겨진 메뉴는 주문할 수 없다.")
-    @Test
-    void createNotDisplayedMenuOrder() {
-        final UUID menuId = menuRepository.save(menu(19_000L, false, menuProduct())).getId();
-        final Order expected = createOrderRequest(OrderType.DELIVERY, "경기도 의왕시", createOrderLineItemRequest(menuId, 19_000L, 3L));
-        assertThatThrownBy(() -> deliveryOrderService.create(expected))
-                .isInstanceOf(IllegalStateException.class);
-    }
-
-    @DisplayName("주문한 메뉴의 가격은 실제 메뉴 가격과 일치해야 한다.")
-    @Test
-    void createNotMatchedMenuPriceOrder() {
-        final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
-        final Order expected = createOrderRequest(OrderType.DELIVERY, "경기도 의왕시", createOrderLineItemRequest(menuId, 16_000L, 3L));
-        assertThatThrownBy(() -> deliveryOrderService.create(expected))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("주문을 접수한다.")
-    @Test
-    void accept() {
-        final UUID orderId = orderRepository.save(order(OrderStatus.WAITING, orderTable(true, 4))).getId();
-        final Order actual = deliveryOrderService.accept(orderId);
-        assertThat(actual.getStatus()).isEqualTo(OrderStatus.ACCEPTED);
-    }
-
-    @DisplayName("접수 대기 중인 주문만 접수할 수 있다.")
-    @EnumSource(value = OrderStatus.class, names = "WAITING", mode = EnumSource.Mode.EXCLUDE)
-    @ParameterizedTest
-    void accept(final OrderStatus status) {
-        final UUID orderId = orderRepository.save(order(status, orderTable(true, 4))).getId();
-        assertThatThrownBy(() -> deliveryOrderService.accept(orderId))
-                .isInstanceOf(IllegalStateException.class);
-    }
-
-    @DisplayName("배달 주문을 접수되면 배달 대행사를 호출한다.")
-    @Test
-    void acceptDeliveryOrder() {
-        final UUID orderId = orderRepository.save(order(OrderStatus.WAITING, "서울시 송파구 위례성대로 2")).getId();
-        final Order actual = deliveryOrderService.accept(orderId);
-        assertAll(
-                () -> assertThat(actual.getStatus()).isEqualTo(OrderStatus.ACCEPTED),
-                () -> assertThat(kitchenridersClient.getOrderId()).isEqualTo(orderId),
-                () -> assertThat(kitchenridersClient.getDeliveryAddress()).isEqualTo("서울시 송파구 위례성대로 2")
-        );
-    }
-
-    @DisplayName("주문을 서빙한다.")
-    @Test
-    void serve() {
-        final UUID orderId = orderRepository.save(order(OrderStatus.ACCEPTED)).getId();
-        final Order actual = deliveryOrderService.serve(orderId);
-        assertThat(actual.getStatus()).isEqualTo(OrderStatus.SERVED);
-    }
-
-    @DisplayName("접수된 주문만 서빙할 수 있다.")
-    @EnumSource(value = OrderStatus.class, names = "ACCEPTED", mode = EnumSource.Mode.EXCLUDE)
-    @ParameterizedTest
-    void serve(final OrderStatus status) {
-        final UUID orderId = orderRepository.save(order(status)).getId();
-        assertThatThrownBy(() -> deliveryOrderService.serve(orderId))
-                .isInstanceOf(IllegalStateException.class);
-    }
-
-    @DisplayName("주문을 배달한다.")
-    @Test
-    void startDelivery() {
-        final UUID orderId = orderRepository.save(order(OrderStatus.SERVED, "서울시 송파구 위례성대로 2")).getId();
-        final Order actual = deliveryOrderService.startDelivery(orderId);
-        assertThat(actual.getStatus()).isEqualTo(OrderStatus.DELIVERING);
-    }
-
-    @DisplayName("배달 주문만 배달할 수 있다.")
-    @Test
-    void startDeliveryWithoutDeliveryOrder() {
-        final UUID orderId = orderRepository.save(order(OrderStatus.SERVED)).getId();
-        assertThatThrownBy(() -> deliveryOrderService.startDelivery(orderId))
-                .isInstanceOf(IllegalStateException.class);
-    }
-
-    @DisplayName("서빙된 주문만 배달할 수 있다.")
-    @EnumSource(value = OrderStatus.class, names = "SERVED", mode = EnumSource.Mode.EXCLUDE)
-    @ParameterizedTest
-    void startDelivery(final OrderStatus status) {
-        final UUID orderId = orderRepository.save(order(status, "서울시 송파구 위례성대로 2")).getId();
-        assertThatThrownBy(() -> deliveryOrderService.startDelivery(orderId))
-                .isInstanceOf(IllegalStateException.class);
-    }
-
-    @DisplayName("주문을 배달 완료한다.")
-    @Test
-    void completeDelivery() {
-        final UUID orderId = orderRepository.save(order(OrderStatus.DELIVERING, "서울시 송파구 위례성대로 2")).getId();
-        final Order actual = deliveryOrderService.completeDelivery(orderId);
-        assertThat(actual.getStatus()).isEqualTo(OrderStatus.DELIVERED);
-    }
-
-    @DisplayName("배달 중인 주문만 배달 완료할 수 있다.")
-    @EnumSource(value = OrderStatus.class, names = "DELIVERING", mode = EnumSource.Mode.EXCLUDE)
-    @ParameterizedTest
-    void completeDelivery(final OrderStatus status) {
-        final UUID orderId = orderRepository.save(order(status, "서울시 송파구 위례성대로 2")).getId();
-        assertThatThrownBy(() -> deliveryOrderService.completeDelivery(orderId))
-                .isInstanceOf(IllegalStateException.class);
-    }
-
-    @DisplayName("주문을 완료한다.")
-    @Test
-    void complete() {
-        final Order expected = orderRepository.save(order(OrderStatus.DELIVERED, "서울시 송파구 위례성대로 2"));
-        final Order actual = deliveryOrderService.complete(expected.getId());
-        assertThat(actual.getStatus()).isEqualTo(OrderStatus.COMPLETED);
-    }
-
-    @DisplayName("배달 주문의 경우 배달 완료된 주문만 완료할 수 있다.")
-    @EnumSource(value = OrderStatus.class, names = "DELIVERED", mode = EnumSource.Mode.EXCLUDE)
-    @ParameterizedTest
-    void completeDeliveryOrder(final OrderStatus status) {
-        final UUID orderId = orderRepository.save(order(status, "서울시 송파구 위례성대로 2")).getId();
-        assertThatThrownBy(() -> deliveryOrderService.complete(orderId))
-                .isInstanceOf(IllegalStateException.class);
-    }
-
-    @DisplayName("주문의 목록을 조회할 수 있다.")
-    @Test
-    void findAll() {
-        final OrderTable orderTable = orderTableRepository.save(orderTable(true, 4));
-        orderRepository.save(order(OrderStatus.SERVED, orderTable));
-        orderRepository.save(order(OrderStatus.DELIVERED, "서울시 송파구 위례성대로 2"));
-        final List<Order> actual = deliveryOrderService.findAll();
-        assertThat(actual).hasSize(2);
-    }
-
-    private Order createOrderRequest(
-            final OrderType type,
-            final String deliveryAddress,
-            final OrderLineItem... orderLineItems
-    ) {
-        return new Order(UUID.randomUUID(), type, OrderStatus.WAITING, LocalDateTime.now(), new OrderLineItems(Arrays.asList(orderLineItems)), deliveryAddress, null, null);
-    }
-
-    private Order createOrderRequest(final OrderType orderType, final OrderLineItem... orderLineItems) {
-        return createOrderRequest(orderType, Arrays.asList(orderLineItems));
-    }
-
-    private Order createOrderRequest(final OrderType orderType, final List<OrderLineItem> orderLineItems) {
-        OrderTable orderTable = orderTable(false, 0);
-        return new Order(UUID.randomUUID(), orderType, OrderStatus.WAITING, LocalDateTime.now(), new OrderLineItems(orderLineItems), null, orderTable, orderTable.getId());
-    }
-
-    private Order createOrderRequest(
-            final OrderType type,
-            final UUID orderTableId,
-            final OrderLineItem... orderLineItems
-    ) {
-        return new Order(UUID.randomUUID(), type, OrderStatus.WAITING, LocalDateTime.now(), new OrderLineItems(Arrays.asList(orderLineItems)), null, null, orderTableId);
-    }
-}
